Add Firefox profiles with granted permissions and fake devices
In order to test calls a camera and a microphone need to be faked by the
acceptance tests. Fortunately, Firefox provides fake video and audio
streams that can be set through settings in the Firefox profile (either
video and audio or only audio, but not only video).

Note, however, that those fake streams can not be controlled; the audio
stream is always playing, so it can not be used to extensively test a
conversation between different participants. It is good enough for basic
tests, though.

Before using the camera or the microphone Firefox asks the user to grant
access to them, even if the fake ones are going to be used. This is done
through a dialog outside the Selenium domain, so it can not be accepted
or declined directly from the acceptance tests.

However, it is possible to disable that dialog through a setting in the
Firefox profile too. When the dialog is disabled the permissions are
automatically granted whenever they are needed without any interaction.

By default Selenium uses a clean Firefox profile when a new session is
started, but it also allows the customization of the profile through a
zipped "user.js" file. The contents of that file have to be provided in
the "firefox_profile" capability when the Firefox session is created.

In the Mink extension for Behat several Mink sessions can be defined in
the "behat.yml" file. Each Mink session uses a different browser session
in Selenium, and each of those browser sessions is initialized with the
capabilities provided in the "behat.yml" file.

From the point of view of the acceptance tests each Mink session is an
actor, so different actors can use different browsers with different
capabilities.

Due to all this several new actors were introduced meant to be used in
tests that require granted media permissions with no camera nor
microphone, granted media permissions with microphone but no camera, and
granted media permissions with microphone and camera.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>