Ignore some operations when the container is hidden
Some operations on the virtual list, like reloading it, updating the
visible elements or scrolling to certain element, require that the
container is visible. Before those operations messed with the state of
the virtual list when called with a hidden container, but now they are
simply ignored.

Adding new elements is still possible while the virtual list is hidden,
but "reload()" must be explicitly called once the container is visible
again for the added elements to be loaded. Like in a regular "reload()",
if the width has not changed there is no need to reload all the
elements; only those that were added while the container was hidden will
be added in that case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>